requirements.txt
pyrogram
TgCrypto
motor
pymongo
dnspython
python-dotenv
aiohttp